OpenAI signs $38bn cloud deal with Amazon
OpenAI will end its partnership with Microsoft and start a new one with Amazon in the future. The decision was made because OpenAI needs more freedom to develop artificial intelligence, and Amazon, being a huge company in computer technology, can also support OpenAI in becoming a leader in AI development.